feat: pr comments fixes
